Check in Solr configs, add suggestion
Suggestion is so far based on ID, which is highly useless. Once we know
the schema and add some example data or something, we'll get a good idea
about what it is we should actually be using. Text search the
`solrconfig.xml` file for "suggest" to find the interesting part from
this week.

Most of the stuff in here was autogenerated by Solr. I've only edited a
few things: the schema (to Tim's example DB) and the data import handler
stuff (with accompanying settings in the `db-data-config.xml` file.

This may all be for nothing though if the DB team is insistant on using
MySQL. This is set up for Postgres and while it should be easy to switch
(get the JDBC, apply settings), it may be much more difficult. Changes
to these files will ensue.